+++ src/main/java/com/damoa/domain/community/dto/CreateCommunityDto.java
@@ -1,8 +1,8 @@
-package com.damoa.web.dto.community;
+package com.damoa.domain.community.dto;
 
-import com.damoa.domain.community.Category;
-import com.damoa.domain.community.Community;
-import com.damoa.domain.community.MemberGrade;
+import com.damoa.domain.community.entity.Category;
+import com.damoa.domain.community.entity.Community;
+import com.damoa.domain.community.entity.MemberGrade;
 import com.damoa.domain.member.Member;
 import lombok.*;
 import org.hibernate.validator.constraints.Length;
