+++ src/main/java/com/damoa/domain/community/entity/Community.java
@@ -1,4 +1,4 @@
-package com.damoa.domain.community;
+package com.damoa.domain.community.entity;
 
 import com.damoa.domain.BaseTimeEntity;
 import com.damoa.domain.member.Member;
