+++ src/main/java/com/damoa/domain/community/entity/MemberGrade.java
@@ -1,4 +1,4 @@
-package com.damoa.domain.community;
+package com.damoa.domain.community.entity;
 
 public enum MemberGrade {
     OWNER, ADMIN, USER
