+++ src/main/java/com/damoa/domain/community/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.damoa.repository;
+package com.damoa.domain.community.repository;
 
-import com.damoa.domain.community.Category;
+import com.damoa.domain.community.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
