+++ src/main/java/com/damoa/domain/community/repository/CommunityRepository.java
@@ -1,6 +1,6 @@
-package com.damoa.repository;
+package com.damoa.domain.community.repository;
 
-import com.damoa.domain.community.Community;
+import com.damoa.domain.community.entity.Community;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
