+++ src/main/java/com/damoa/domain/community/repository/SearchCommunityRepository.java
@@ -1,6 +1,6 @@
-package com.damoa.repository;
+package com.damoa.domain.community.repository;
 
-import com.damoa.domain.community.Community;
+import com.damoa.domain.community.entity.Community;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
