+++ src/main/java/com/damoa/domain/post/Post.java
@@ -1,7 +1,7 @@
 package com.damoa.domain.post;
 
 import com.damoa.domain.BaseTimeEntity;
-import com.damoa.domain.community.Community;
+import com.damoa.domain.community.entity.Community;
 import com.damoa.domain.member.Member;
 import lombok.*;
