+++ src/main/java/com/damoa/web/controller/MemberController.java
@@ -3,11 +3,11 @@ package com.damoa.web.controller;
 import com.damoa.common.PageableValidator;
 import com.damoa.domain.member.Member;
 import com.damoa.service.AuthService;
-import com.damoa.service.CommunityService;
+import com.damoa.domain.community.service.CommunityService;
 import com.damoa.service.MemberService;
 import com.damoa.service.PostService;
 import com.damoa.web.dto.common.CurrentMember;
-import com.damoa.web.dto.community.JoinedCommunityDto;
+import com.damoa.domain.community.dto.JoinedCommunityDto;
 import com.damoa.web.dto.member.*;
 import com.damoa.web.dto.post.PostDto;
 import lombok.RequiredArgsConstructor;
