+++ src/main/java/com/damoa/web/dto/post/PostDto.java
@@ -1,7 +1,7 @@
 package com.damoa.web.dto.post;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.damoa.domain.community.Community;
+import com.damoa.domain.community.entity.Community;
 import com.damoa.domain.member.Member;
 import com.damoa.domain.post.Post;
 import lombok.Getter;
