+++ src/test/java/com/damoa/service/PostServiceTest.java
@@ -1,14 +1,14 @@
 package com.damoa.service;
 
-import com.damoa.domain.community.Category;
-import com.damoa.domain.community.Community;
-import com.damoa.domain.community.CommunityMember;
-import com.damoa.domain.community.MemberGrade;
+import com.damoa.domain.community.entity.Category;
+import com.damoa.domain.community.entity.Community;
+import com.damoa.domain.community.entity.CommunityMember;
+import com.damoa.domain.community.entity.MemberGrade;
 import com.damoa.domain.member.Member;
 import com.damoa.domain.post.Comment;
 import com.damoa.domain.post.Post;
 import com.damoa.repository.CommentRepository;
-import com.damoa.repository.CommunityMemberRepository;
+import com.damoa.domain.community.repository.CommunityMemberRepository;
 import com.damoa.repository.ImageRepository;
 import com.damoa.repository.PostRepository;
 import com.damoa.web.dto.post.CreatePostDto;
