+++ src/test/java/com/damoa/web/controller/MemberControllerTest.java
@@ -2,8 +2,8 @@ package com.damoa.web.controller;
 
 import com.damoa.domain.member.Member;
 import com.damoa.web.common.BaseControllerTest;
-import com.damoa.web.dto.community.CommunityDto;
-import com.damoa.web.dto.community.CreateCommunityDto;
+import com.damoa.domain.community.dto.CommunityDto;
+import com.damoa.domain.community.dto.CreateCommunityDto;
 import com.damoa.web.dto.member.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
