+++ src/test/java/com/damoa/web/controller/PostControllerTest.java
@@ -2,7 +2,7 @@ package com.damoa.web.controller;
 
 import com.damoa.domain.member.Member;
 import com.damoa.web.common.BaseControllerTest;
-import com.damoa.web.dto.community.CreateCommunityDto;
+import com.damoa.domain.community.dto.CreateCommunityDto;
 import com.damoa.web.dto.member.JwtTokenDto;
 import com.damoa.web.dto.member.LoginMemberDto;
 import com.damoa.web.dto.post.CreatePostDto;
